+++ src/writer/Writer.hh
@@ -72,7 +72,7 @@ protected:
 	const BoundarySize boundarySize;
 
     //! dimensions of the grid in x- and y-direction.
-    const unsigned int nX, nY;
+    const size_t nX, nY;
 
     //! current time step
     size_t timeStep;
